changed iterator & counter type

+++ src/sage/matrix/matrix2.pyx
@@ -10244,7 +10244,7 @@ cdef class Matrix(Matrix1):
             sage: a.density()
             0
         """
-        cdef int x, y, k
+        cdef Py_ssize_t x, y, k
         k = 0
         nr = self.nrows()
         nc = self.ncols()
